added sass and light refactor
