+++ src/index.js
@@ -1,22 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Router } from 'react-router-dom';
-// import './index.css';
-// import App from './App';
+
+import './styles/main.scss';
+import App from './App';
 // import reportWebVitals from './reportWebVitals';
 
-import { Routes } from './routes/Routes';
-
-import { createBrowserHistory } from 'history';
-
-const history = createBrowserHistory();
-
-ReactDOM.render(
-    <Router history={history}>
-        <Routes />
-    </Router>,
-    document.getElementById('root')
-);
+ReactDOM.render(<App />, document.getElementById('root'));
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
