+++ src/styles/_globals.scss
@@ -1,0 +1,5 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
