+++ src/styles/main.scss
@@ -1,0 +1,1 @@
+@import './globals';
