+++ src/utils/cookiesManager.js
@@ -2,7 +2,7 @@ import Cookies from 'js-cookie';
 import { USER_SESSION } from '../constants/cookie_const';
 
 export const setCookie = (data, cookie = USER_SESSION) => {
-    Cookies.set(cookie, data, { expires: 1 });
+    Cookies.set(cookie, JSON.stringify(data), { expires: 1 });
 };
 
 export const getCookie = (cookie = USER_SESSION) => {
